refactor #6 - extract out button handlers
