+++ judgments/views/__init__.py
@@ -1,6 +1,4 @@
-import json
 import math
-from typing import Optional, Union
 
 import environ
 from caselawclient.Client import (
@@ -9,9 +7,7 @@ from caselawclient.Client import (
     MarklogicResourceNotFoundError,
     api_client,
 )
-from django.contrib import messages
-from django.http import Http404, HttpResponse, HttpResponseBadRequest
-from django.shortcuts import redirect
+from django.http import Http404, HttpResponse
 from django.template import loader
 from django.utils.translation import gettext
 from requests_toolbelt.multipart import decoder
@@ -19,6 +15,11 @@ from requests_toolbelt.multipart import decoder
 from judgments.models import SearchResult, SearchResults
 from judgments.utils import create_s3_client
 
+from .button_handlers import (  # noqa
+    assign_judgment_button,
+    hold_judgment_button,
+    prioritise_judgment_button,
+)
 from .delete import delete  # noqa
 from .detail import detail  # noqa
 from .detail_xml import detail_xml  # noqa
@@ -27,57 +28,6 @@ from .edit_judgment import EditJudgmentView  # noqa
 env = environ.Env()
 akn_namespace = {"akn": "http://docs.oasis-open.org/legaldocml/ns/akn/3.0"}
 uk_namespace = {"uk": "https://caselaw.nationalarchives.gov.uk/akn"}
-
-
-def hold_judgment_button(request):
-    judgment_uri = request.POST["judgment_uri"]
-    # we probably shouldn't hold if the judgment isn't assigned but we won't check
-    hold = request.POST["hold"]
-    if hold not in ["false", "true"]:
-        raise RuntimeError("Hold value must be '0' or '1'")
-    api_client.set_property(judgment_uri, "editor-hold", hold)
-    target_uri = request.META.get("HTTP_REFERER") or "/"
-    if hold == "true":
-        word = "held"
-    else:
-        word = "released"
-    messages.success(request, f"Judgment {word}.")
-    return redirect(target_uri)
-
-
-def assign_judgment_button(request):
-    judgment_uri = request.POST["judgment_uri"]
-    api_client.set_property(judgment_uri, "assigned-to", request.user.username)
-    target_uri = request.META.get("HTTP_REFERER") or "/"
-    if request.headers.get("X-Requested-With") == "XMLHttpRequest":
-        return HttpResponse(
-            json.dumps({"assigned_to": request.user.username}),
-            content_type="application/json",
-        )
-    else:
-        messages.success(request, "Judgment assigned to you.")
-        return redirect(target_uri)
-
-
-def prioritise_judgment_button(request):
-    """Editors can let other editors know that some judgments are more important than others."""
-
-    def parse_priority(priority: Union[str, int]) -> Optional[str]:
-        # note: use 09 if using numbers less than 10.
-        priorities = {"low": "10", "medium": "20", "high": "30"}
-        priority_string = priority.lower().strip()
-        return priorities.get(priority_string)
-
-    judgment_uri = request.POST["judgment_uri"]
-    priority = parse_priority(request.POST["priority"])
-    if priority:
-        api_client.set_property(judgment_uri, "editor-priority", priority)
-        target_uri = request.META.get("HTTP_REFERER") or "/"
-
-        messages.success(request, "Judgment priority set.")
-        return redirect(target_uri)
-
-    return HttpResponseBadRequest("Priority string not recognised")
 
 
 def index(request):
